# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI & ML Assessment API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DataPoint {
  x: number;
  y: number;
}

export interface RegressionInput {
  points: DataPoint[];
}

export interface RegressionResult {
  slope: number;
  intercept: number;
  rSquared: number;
  predictions: DataPoint[];
}

export type ClassificationInputTrainingPointsItem = {
  x: number;
  y: number;
  label: string;
};

export interface ClassificationInput {
  trainingPoints: ClassificationInputTrainingPointsItem[];
  queryPoint: DataPoint;
  k: number;
}

export type ClassificationResultNeighborsItem = {
  x: number;
  y: number;
  label: string;
  distance: number;
};

export interface ClassificationResult {
  predictedLabel: string;
  neighbors: ClassificationResultNeighborsItem[];
}

export interface StatisticsInput {
  data: number[];
}

export interface StatisticsResult {
  mean: number;
  median: number;
  mode?: number | null;
  stdDev: number;
  variance: number;
  min: number;
  max: number;
  count: number;
}

export interface KMeansInput {
  points: DataPoint[];
  k: number;
  maxIterations?: number;
}

export interface ClusteredPoint {
  x: number;
  y: number;
  cluster: number;
}

export interface Centroid {
  x: number;
  y: number;
  cluster: number;
}

export interface KMeansResult {
  clusters: ClusteredPoint[];
  centroids: Centroid[];
  iterations: number;
}