# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI & ML Assessment API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ml/statistics:
    post:
      operationId: computeStatistics
      tags: [ml]
      summary: Compute Statistics
      description: Compute descriptive statistics for a dataset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatisticsInput"
      responses:
        "200":
          description: Statistics result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatisticsResult"
...
components:
  schemas:
...
    StatisticsInput:
      type: object
      properties:
        data:
          type: array
          items:
            type: number
      required:
        - data
    StatisticsResult:
      type: object
      properties:
        mean:
          type: number
        median:
          type: number
        mode:
          type: number
          nullable: true
        stdDev:
          type: number
        variance:
          type: number
        min:
          type: number
        max:
          type: number
        count:
          type: integer
      required:
        - mean
        - median
        - stdDev
        - variance
        - min
        - max
        - count
...