# lib/api-zod/src/generated/types/statisticsResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI & ML Assessment API
 * OpenAPI spec version: 0.1.0
 */

export interface StatisticsResult {
  mean: number;
  median: number;
  mode?: number | null;
  stdDev: number;
  variance: number;
  min: number;
  max: number;
  count: number;
}